pyproject.toml
[tool.poetry]
name = "playlist-bot"
version = "0.1.0"
description = ""
authors = ["MaHryCT3 <mahryct123@gmail.com>"]
readme = "README.md"
packages = [{include = "playlist_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.22.2"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
pydantic = "^1.10.2"
asyncpg = "^0.26.0"
loguru = "^0.6.0"
pendulum = "^2.1.2"
APScheduler = "^3.9.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
python-dotenv = "^0.21.0"
flake8-annotations = "^2.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 99
exclude = '(\.git|\.venv|\.vscode)'